1)change maping pojo 2)add UserOperationService 3)change length password in user on 80 char

+++ src/main/java/ru/bpr/onlinestore/portal/services/loading/CategoryOperationService.java
@@ -1,15 +1,15 @@
 package ru.bpr.onlinestore.portal.services.loading;
 
 import ru.bpr.onlinestore.portal.services.models.Category;
-import ru.bpr.onlinestore.portal.services.models.Offer;
 
-import java.lang.Object;import java.util.List;
+import java.math.BigInteger;
+import java.util.List;
 
 public interface CategoryOperationService
 {
     List<Category> loadCategories();
 
-    Category loadById(int id);
+    Category loadById(BigInteger id);
 
     void add(Category category);
