+++ src/main/java/ru/bpr/onlinestore/portal/services/loading/OfferOperationService.java
@@ -2,13 +2,14 @@ package ru.bpr.onlinestore.portal.services.loading;
 
 import ru.bpr.onlinestore.portal.services.models.Offer;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface OfferOperationService
 {
     List<Offer> loadOffers();
 
-    Offer loadById(int id);
+    Offer loadById(BigInteger id);
 
     void add(Offer offer);
